Update day04_p1.py
Found what the error was. I had misunderstood how to think if there are two similar numbers next to each other. If there are three identical numbers next to each other, there must be another number that is next to each other, otherwise it is not valid.

Added some code to check how many times a number is repeted. Now I get that the number of valid inputs are to low.